+++ script/cpp/src/tvgw_converter.cpp
@@ -1,5 +1,4 @@
-#include"tvgw.h"
-#include"tvgw_converter.h"
+#include"tvgw/all.h"
 
 extern "C" void tvgw_ConvertToFormatChunk(TVGW_File_Chunk* chunk, TVGW_Format_Contents* fmt){
     // fmt[space]
