+++ script/cpp/src/tvgw_filewrite.cpp
@@ -1,7 +1,7 @@
 #include<vector>
 #include<fstream>
 
-#include"tvgw_filewrite.h"
+#include"tvgw/all.h"
 
 extern "C" int tvgw_WriteFile(const char* filename, TVGW_RIFF_File* filedata){
     std::ofstream fout(filename, std::ios::binary);
